backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
httpx==0.28.1
python-dotenv==1.1.1
